# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klub Pływacki Jedynka API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsArticle {
  id: number;
  title: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  publishedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNewsArticle {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  excerpt: string;
  imageUrl?: string | null;
  publishedAt: string;
}

export interface ErrorResponse {
  error: string;
}