# lib/api-zod/src/generated/types/createNewsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klub Pływacki Jedynka API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNewsArticle {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  excerpt: string;
  imageUrl?: string | null;
  publishedAt: Date;
}