# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klub Pływacki Jedynka API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: number;
  title: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  publishedAt: Date;
  createdAt: Date;
  updatedAt: Date;
}